refactoring after adding support for k8s
