+++ src/core/src/main/scala/gwi/mawex/master/MasterCmd.scala
@@ -1,0 +1,22 @@
+package gwi.mawex.master
+
+import gwi.mawex.ClusterService
+import org.backuity.clist.{Command, opt}
+import scala.concurrent.duration._
+
+import scala.concurrent.{Await, ExecutionContext}
+
+object MasterCmd extends Command(name = "master", description = "launches master") with ClusterService {
+  import ClusterService._
+
+  var progressingTaskTimeout  = opt[Int](useEnv = true, default = 60*60, description = "timeout for a task progression in seconds")
+  var pendingTaskTimeout      = opt[Int](useEnv = true, default = 3*24, description = "timeout for a pending task in hours")
+  var masterId                = opt[String](useEnv = true, default = "master", name="master-id", description = "Unique identifier of this master node")
+
+  def run(): Unit = {
+    val system = buildClusterSystem(hostAddress, seedNodes, seedNodes.size)
+    clusterSingletonActorRef(Master.Config(masterId, progressingTaskTimeout.seconds, pendingTaskTimeout.hours), system)
+    system.whenTerminated.onComplete(_ => System.exit(0))(ExecutionContext.Implicits.global)
+    sys.addShutdownHook(Await.result(system.terminate(), 10.seconds))
+  }
+}
